Fixed a build error (conversion to ULONG) in fuseio.c to allow x64 version to build. TortoiseSVN screwed up "sources" by overwriting it with the output from a diff command so I undid its handiwork.
git-svn-id: file:///home/konrad/curqtr/cse481a/schooldir/repo@69 0ad51bf3-3eb5-41bf-9b2c-be99ea3c4cf2

+++ ifs/fuse/wxp/fuseio.c
@@ -76,7 +76,7 @@ FuseAddUserspaceIrp (
     if(!BackslashPosition) {
         ModuleName = FileName;
     } else {
-        ULONG ModuleNameLength = BackslashPosition - FileName;
+        ULONG ModuleNameLength = (ULONG)(BackslashPosition - FileName);
         ModuleName = (WCHAR*) ExAllocatePool(PagedPool, sizeof(WCHAR) * (ModuleNameLength + 1));
         memcpy(ModuleName, FileName, sizeof(WCHAR) * (ModuleNameLength));
         ModuleName[ModuleNameLength] = '\0';
